Using 32 bit blocks instead of 64 bit to store number

+++ ubint.h
@@ -4,7 +4,7 @@
 #include <stdint.h>
 
 typedef struct unbounded_integer {
-	uint64_t *blocks;
+	uint32_t *blocks;
 	uint64_t n_blocks;
 	uint8_t sign;
 } ubint;
